Created seperate execution function class file
