+++ classes/main.rb
@@ -1,4 +1,5 @@
 require './classes/app'
+require './classes/execute_option'
 app = App.new
 
 def main(app)
@@ -20,21 +21,21 @@ def main(app)
   end
 end
 
-def execute_option(app, number)
-  case number
-  when 1
-    app.list_books
-  when 2
-    app.list_people
-  when 3
-    app.create_person
-  when 4
-    app.create_book
-  when 5
-    app.create_rental
-  when 6
-    app.list_rentals
-  end
-end
+# def execute_option(app, number)
+#   case number
+#   when 1
+#     app.list_books
+#   when 2
+#     app.list_people
+#   when 3
+#     app.create_person
+#   when 4
+#     app.create_book
+#   when 5
+#     app.create_rental
+#   when 6
+#     app.list_rentals
+#   end
+# end
 
 main(app)
